Fixed references to the top container uri
The record mapper was trying to pull the top container uri from a location
where that field does not exist. Instead, that field is called "ref" at that
location. That reference has been renamed (and points to) the ref field. The
uri field has been sourced from the _resolved top container, where that field
does exist.