Enhancement: Pull in instance information from a parent object (#25)
Added a separate method for getting the top container instances from a record. Digital object instances are still ignored from this list.

- If the record has top container instances, then those will be used.
- If the record does not have any top container instances, then the method will walk up the record's resource tree (checking to see if the record has a "parent" or a "resource") and will look for a record that does have top container instances. If one is found, then the top container instance information will be pulled from that record.
- If no top container instances are found for the current record, nor for any of the ancestors of the current record, then the mapper declares that the current record does not have any top container instances, and will continue on as normal.

This new functionality does not add any new settings, but does slightly change the functionality of the `:requests_permitted_for_containers_only` setting. If the current record does not have instance information, but a parent record does, and this setting is active, then the record will be requestable.