Added accessible hover text to the "Aeon Request" button, when disabled (#23)
When the Aeon Request button is disabled, the button will be surrounded with an invisible and tab-able div, that displays a tooltip on tab and on hover. The message of this tooltip is configurable in the locales. The button will also have a "visually hidden" span that contains the same text as tooltip, allowing screen readers to find and say the message.