Update Readme.md
Added additional explanation for Aeon configurations that need to be completed before the plugin can be used or tested. Altered the code for enabling the plugin in order to match the ArchivesSpace best practices and also added some instructions to provide additional clarity. Updated an outdated link to Aeon OpenURL documentation in the Atlas Support documentation.